make mempool depend on ShardChainStorage
